Dynamic camera working

+++ code/src/components/SimpleComponents.hpp
@@ -47,14 +47,14 @@ class Camera : public Component{
   virtual ~Camera(){};
   virtual glm::vec3 getViewDir() = 0;
   virtual glm::mat4 getView() = 0;
-  virtual glm::mat4 getPerspective(double aspect) = 0;
+  virtual glm::mat4 getPerspective(float aspect) = 0;
 
 };
 
 class StaticCamera : public Camera{
  public:
   StaticCamera() : fov(45.0), near(.01), far(100.0), pose(glm::vec3(0.0)), lookat(glm::vec3(0.0)), updir(glm::vec3(0.0, 1.0, 0.0)){}
-  StaticCamera(double fov, const glm::vec3 &loc, const glm::vec3 &look) : fov(fov), near(.01), far(100.0), pose(loc), lookat(look), updir(glm::vec3(0.0, -1.0, 0.0)){}
+  StaticCamera(float fov, const glm::vec3 &loc, const glm::vec3 &look) : fov(fov), near(.01), far(100.0), pose(loc), lookat(look), updir(glm::vec3(0.0, -1.0, 0.0)){}
 
   glm::vec3 getViewDir(){return(lookat-pose.loc);}
 
@@ -66,13 +66,13 @@ class StaticCamera : public Camera{
     ));
   }
 
-  glm::mat4 getPerspective(double aspect){
+  glm::mat4 getPerspective(float aspect){
     return(glm::perspective(fov, aspect, near, far));
   }
 
-  double fov;
-  double near;
-  double far;
+  float fov;
+  float near;
+  float far;
 
   Pose pose;
   glm::vec3 lookat;
